+++ include/metashell/empty_environment.hpp
@@ -1,8 +1,8 @@
-#ifndef METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
-#define METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
+#ifndef METASHELL_EMPTY_ENVIRONMENT_HPP
+#define METASHELL_EMPTY_ENVIRONMENT_HPP
 
 // Metashell - Interactive C++ template metaprogramming shell
-// Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
+// Copyright (C) 2017, Abel Sinkovics (abel@sinkovics.hu)
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -17,26 +17,25 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/data/includes.hpp>
-#include <metashell/iface/header_discoverer.hpp>
+#include <metashell/iface/environment.hpp>
 
 namespace metashell
 {
-  class header_discoverer_constant : public iface::header_discoverer
+  class empty_environment : public iface::environment
   {
   public:
-    header_discoverer_constant(
-        std::vector<boost::filesystem::path> sysincludes_,
-        std::vector<boost::filesystem::path> quoteincludes_);
+    explicit empty_environment(const boost::filesystem::path& internal_dir_);
 
-    virtual std::vector<boost::filesystem::path>
-    include_path(data::include_type type_) override;
+    virtual void append(const std::string& s_) override;
+    virtual std::string get() const override;
+    virtual std::string get_appended(const std::string& s_) const override;
 
-    virtual std::set<boost::filesystem::path>
-    files_included_by(const std::string&) override;
+    virtual const data::headers& get_headers() const override;
+
+    virtual std::string get_all() const override;
 
   private:
-    data::includes _includes;
+    data::headers _headers;
   };
 }