+++ include/metashell/header_discoverer_wave.hpp
@@ -1,8 +1,8 @@
-#ifndef METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
-#define METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
+#ifndef METASHELL_HEADER_DISCOVERER_WAVE_HPP
+#define METASHELL_HEADER_DISCOVERER_WAVE_HPP
 
 // Metashell - Interactive C++ template metaprogramming shell
-// Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
+// Copyright (C) 2017, Abel Sinkovics (abel@sinkovics.hu)
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -17,26 +17,25 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/data/includes.hpp>
 #include <metashell/iface/header_discoverer.hpp>
+
+#include <metashell/data/wave_config.hpp>
 
 namespace metashell
 {
-  class header_discoverer_constant : public iface::header_discoverer
+  class header_discoverer_wave : public iface::header_discoverer
   {
   public:
-    header_discoverer_constant(
-        std::vector<boost::filesystem::path> sysincludes_,
-        std::vector<boost::filesystem::path> quoteincludes_);
+    explicit header_discoverer_wave(data::wave_config config_);
 
     virtual std::vector<boost::filesystem::path>
     include_path(data::include_type type_) override;
 
     virtual std::set<boost::filesystem::path>
-    files_included_by(const std::string&) override;
+    files_included_by(const std::string& exp_) override;
 
   private:
-    data::includes _includes;
+    data::wave_config _config;
   };
 }