+++ lib/core/empty_environment.cpp
@@ -1,8 +1,5 @@
-#ifndef METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
-#define METASHELL_HEADER_DISCOVERER_CONSTANT_HPP
-
 // Metashell - Interactive C++ template metaprogramming shell
-// Copyright (C) 2016, Abel Sinkovics (abel@sinkovics.hu)
+// Copyright (C) 2017, Abel Sinkovics (abel@sinkovics.hu)
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -17,27 +14,32 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-#include <metashell/data/includes.hpp>
-#include <metashell/iface/header_discoverer.hpp>
+#include <metashell/empty_environment.hpp>
 
 namespace metashell
 {
-  class header_discoverer_constant : public iface::header_discoverer
+  empty_environment::empty_environment(
+      const boost::filesystem::path& internal_dir_)
+    : _headers(internal_dir_)
   {
-  public:
-    header_discoverer_constant(
-        std::vector<boost::filesystem::path> sysincludes_,
-        std::vector<boost::filesystem::path> quoteincludes_);
+  }
 
-    virtual std::vector<boost::filesystem::path>
-    include_path(data::include_type type_) override;
+  void empty_environment::append(const std::string&)
+  {
+    // ignore
+  }
 
-    virtual std::set<boost::filesystem::path>
-    files_included_by(const std::string&) override;
+  std::string empty_environment::get() const { return ""; }
 
-  private:
-    data::includes _includes;
-  };
+  std::string empty_environment::get_appended(const std::string&) const
+  {
+    return "";
+  }
+
+  const data::headers& empty_environment::get_headers() const
+  {
+    return _headers;
+  }
+
+  std::string empty_environment::get_all() const { return ""; }
 }
-
-#endif