+++ lib/core/header_discoverer_vc.cpp
@@ -93,7 +93,7 @@ namespace metashell
   std::vector<boost::filesystem::path>
   header_discoverer_vc::include_path(data::include_type type_)
   {
-    return type_ == data::include_type::sys ? _includes.sys : _includes.quote;
+    return get(type_, _includes);
   }
 
   std::set<boost::filesystem::path>